+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ###### July 17, 2022
 
-- Implemention struacture application
-- Implemention Configuration Based On [config.json]
-- Add Statics Assets
+- Implementation struacture application
+- Implementation configuration based on [config.json]
+- Add statics assets
+
+###### July 18, 2022
+
+- Edit [config.json] file
+- Create hooks [useVideoInfo] for get duration video by using [config.json]
+- Arrange dynamically sequence between videos by using [useVideoInfo]
+- Implementation audio in background
+- Mute all videos
+- Implementation simple animations & style to videos